Refactor data-cleaning code
Until now, every data-cleaning procedure mapped the change to the entire
review dataset, meaning that the whole cleaning process was a series of
mapping functions. This introduced code repetition (a `list(map())`
inside every procedure), which in turn made the code more complicated
and harder to refactor.

Now, instead of adressing the whole list of reviews, each procedure
works with a single review. That way each procedure contains only the
code relevant to the actual change it makes to the review and is free
of list-maintaing boilerplate.
The actual mapping is now performed in the `clean` function itself,
using the `pipe_map` utility function (converting the previous
series-of-maps to a map-of-a-series).