Fix model saving/loading issues
When attempting to load saved model's state, error accured if the saved
model's LSTM had more than one layer. This was caused by initialising
the model with the default arguments (i.e. only one LSTM layer) and
then attempting to load its state with multiple layers, causing a
missmatch.

In order to prevent this issue, alongside model's `state_dict`, now the
hyperparameters are also saved, so they can be used to initialise model
before loading it with the saved state.